Much nicer design in the dynamic fragment: adding padding with the element above adding separator between author and image the page numbering has been changed: textview in the layout that appears below the linearLayout
** Whole hierarchy changed **
Scrollview -> ConstraintLayout -> Linear Layout
Now the fragment is scrollable

** Images**
The image is set to take 30% of the screen height and the width of the linear layout!
Very nice

Article is null safe but the fiels can be empty.
We add article fields to the linear layout only if they are not empty